Improve upload file security

+++ smwWeb/views.py
@@ -12,6 +12,7 @@ from wsgiref.util import FileWrapper
 from django.conf import settings
 from django.http import FileResponse
 
+import yaml
 
 from smwWeb.forms import LoginForm, SigninForm, SettingsFileForm
 from smwWeb.models import Account
@@ -103,17 +104,36 @@ def upload_settings(request):
     if request.method == "POST":
         form = SettingsFileForm(request.POST, request.FILES)
         if form.is_valid():
-            dest = path.join(settings.MEDIA_ROOT, "settings", "config_" + request.user.username + ".yml")
+            upload_file = request.FILES['settings']
+            filename_ok = upload_file.name == 'config.yml'
+            content_type_ok = upload_file.content_type == 'application/x-yaml'
+            size_ok = upload_file.size < 1000
+            if filename_ok and content_type_ok and size_ok:
+                for line in upload_file:
+                open_file = upload_file.open()
+                config = yaml.load(open_file)
+                nb_item = 0
+                valid_items = ['timedelta', 'extention', 'advanced', 'delicate_dirs', 'safe_dir', 'filename', 'dirpath', 'external_path', 'dirname', 'local_path']
+                for item in config:
+                    plus = 1 if item in valid_items else -1
+                    nb_item += plus
+                content_ok = nb_item == len(valid_items)
+                if content_ok:
+                    dest = path.join(settings.MEDIA_ROOT, "settings", "config_" + request.user.username + ".yml")
 
-            if not default_storage.exists(dest):
-                default_storage.save(dest, File(request.FILES["settings"]))
+                    if not default_storage.exists(dest):
+                        default_storage.save(dest, File(request.FILES["settings"]))
+                    else:
+                        default_storage.delete(dest)
+                        default_storage.save(dest, File(request.FILES["settings"]))
+
+                    request.user.account.upload_datetime = datetime.datetime.now()
+                    request.user.account.save()
+                    return redirect(member_account)
+                else:
+                    error = True
             else:
-                default_storage.delete(dest)
-                default_storage.save(dest, File(request.FILES["settings"]))
-
-            request.user.account.upload_datetime = datetime.datetime.now()
-            request.user.account.save()
-            return redirect(member_account)
+                error = True
         else:
             error = True
     else:
